% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{bottomUpRecursive}
\alias{bottomUpRecursive}
\title{bottomUpRecursive}
\usage{
bottomUpRecursive(l, p0)
}
\arguments{
\item{l}{return value of recursivePackageDependenciesWrapper()}

\item{p0}{list of those packages whose R code has been modified by the developer}
}
\value{
returns a list of the original query packages plus the packages that directly import them
}
\description{
given a list of packages, determine which packages recursively import
the packages in the list
}
\examples{
\dontrun{
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
l<-recursivePackageDependenciesWrapper(packs)
bur<-bottomUpRecursive(l,c("iterationDriver"))
}

}
