% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{recursivePackageDependenciesWrapper}
\alias{recursivePackageDependenciesWrapper}
\title{recursivePackageDependenciesWrapper}
\usage{
recursivePackageDependenciesWrapper(packs)
}
\arguments{
\item{packs}{list of package names}
}
\value{
returns a list of lists. Each component is indexed by a package name,
and contains a list of package names that are imported by the index package name
}
\description{
determine which packages in packs list have length zero dependencies
}
\details{
the return value l is like:

$cardUtils

character(0) [i.e., cardUtils has no further dependencies within packs list]

$editDriver

[1] "cardUtils"         "clickableImageMap" "heartsCIM"         "logos"             "probTab"
}
\examples{
\dontrun{
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
l<-recursivePackageDependenciesWrapper(packs)
}

}
