% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{retrieveLeafNodes}
\alias{retrieveLeafNodes}
\title{retrieveLeafNodes}
\usage{
retrieveLeafNodes(master)
}
\arguments{
\item{master}{return value of sortedInputForCheckBuildInstallSourcePackageDriver()}
}
\value{
returns a list of the packages in the correct order for processing by
checkBuildInstallSourcePackage()
}
\description{
compute a list of the packages in the correct order for processing by
checkBuildInstallSourcePackage()
}
\details{
the master list may contain some packages that do not need to be processed by
checkBuildInstallSourcePackage(). These are weeded out by sortedInputForcheckBuildInstallSourcePackage()
}
\examples{
\dontrun{
# you need to specify dir, packs that are on your own computer !!
dir1<-"~/personal/hearts/hearts_card_game_bayesian_inference"
dir2<-"packages/inference_packages/inference_packages/"
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
master<-PackageDependencies(sprintf("\%s/\%s",dir1,dir2),packs,vector("list",length(packs)),1,TRUE)
retrieve<-retrieveLeafNodes(master)
}

}
