% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{retrieveNamespace}
\alias{retrieveNamespace}
\title{retrieveNamespace}
\usage{
retrieveNamespace(dir, packs)
}
\arguments{
\item{dir}{character string containing the name of the directory holding packs}

\item{packs}{list of package names}
}
\value{
returns a list containing the intersection of
(1) imported package names and (2) packs list
}
\description{
retrieve a list of the imported packages in a NAMESPACE FILE
}
\examples{
\dontrun{
# you need to specify dir, packs that are on your own computer !!
dir1<-"~/personal/hearts/hearts_card_game_bayesian_inference"
dir2<-"packages/inference_packages/inference_packages/"
dir<-sprintf("\%s/\%s",dir1,dir2)
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
rns<-retrieveNamespace(dir,packs)
}

}
