% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{reversePackageDependencies}
\alias{reversePackageDependencies}
\title{reversePackageDependencies}
\usage{
reversePackageDependencies(l)
}
\arguments{
\item{l}{return value of recursivePackageDependenciesWrapper()}
}
\value{
returns a list whose components are 2 lists:
\item{$original}{
a list whose components are lists of package names that have non-zero length import dependencies
ll$original is same as l, but deleting zero-length elements i.e., leaf nodes
}
\item{$zeros}{
a list whose components are lists of package names that have zero length import dependencies (i.e., leaf nodes)
ll$zeros zero-length elements, leaf nodes that had been deleted in ll$original
}
}
\description{
separate the packages in packs list having length zero or non-zero dependencies
}
\details{
the return value ll is like:

$original [cardUtils is no longer a name of ll$original since it had length 0]

$original$editDriver

[1] "cardUtils"         "clickableImageMap" "heartsCIM"         "logos"             "probTab"

$zeros

$zeros$cardUtils cardUtils is an element of ll$zeros since l[["cardUtils"]] has length 0

[1] "cardUtils"
}
\examples{
\dontrun{
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
l<-recursivePackageDependenciesWrapper(packs)
ll<-reversePackageDependencies(l)
}

}
