% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{sortedInputForCheckBuildInstallSourcePackage}
\alias{sortedInputForCheckBuildInstallSourcePackage}
\title{sortedInputForCheckBuildInstallSourcePackage}
\usage{
sortedInputForCheckBuildInstallSourcePackage(retrieve, burd)
}
\arguments{
\item{retrieve}{return value of retrieveLeafNodes()}

\item{burd}{return value of bottomUpRecursiveDriver()}
}
\value{
returns a list of packages in the correct order to input to
checkBuildInstallSourcePackage()
}
\description{
compute a list of packages in the correct order to input to
checkBuildInstallSourcePackage()
}
\examples{
\dontrun{
# you need to specify dir, packs that are on your own computer !!
dir1<-"~/personal/hearts/hearts_card_game_bayesian_inference"
dir2<-"packages/inference_packages/inference_packages/"
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
master<-PackageDependencies(sprintf("\%s/\%s",dir1,dir2),packs,vector("list",length(packs)),1,TRUE)
l<-retrieveNamespace(sprintf("\%s/\%s",dir1,dir2),packs)
burd<-bottomUpRecursiveDriver(l,c("iterationDriver"),TRUE)
retrieve<-retrieveLeafNodes(master)
s<-sortedInputForCheckBuildInstallSourcePackage(retrieve,burd)
}

}
