% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{bottomUpRecursiveDriver}
\alias{bottomUpRecursiveDriver}
\title{bottomUpRecursiveDriver}
\usage{
bottomUpRecursiveDriver(l, p0, verbose)
}
\arguments{
\item{l}{return value of retrieveNamespace()}

\item{p0}{list of those packages whose R code has been modified by the developer}

\item{verbose}{if TRUE print line indicating the recursion level}
}
\value{
returns a list of all the packages that either directly or indirectly imports the original query packages
}
\description{
compute a list of all the packages that either directly or indirectly import the original query packages
}
\examples{
\dontrun{
# you need to specify dir, packs that are on your own computer !!
dir1<-"~/personal/hearts/hearts_card_game_bayesian_inference"
dir2<-"packages/inference_packages/inference_packages/"
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
l<-retrieveNamespace(sprintf("\%s/\%s",dir1,dir2),packs)
burd<-bottomUpRecursiveDriver(l,c("iterationDriver"),TRUE)
}

}
