% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoviceDeveloperResources2.R
\name{PackageDependencies}
\alias{PackageDependencies}
\title{PackageDependencies}
\usage{
PackageDependencies(packs, master, n, verbose)
}
\arguments{
\item{packs}{list of package names}

\item{master}{list whose componenets are lists indexed by integer recursion level
the components of each recursion level are the return values of
recursivePackageDependenciesWrapper() and reversePackageDependencies()}

\item{n}{integer recursion level}

\item{verbose}{if TRUE print line indicating the recursion level}
}
\value{
returns a list whose components are lists indexed by the integer recursion level:
\item{l}{
return value of recursivePackageDependenciesWrapper()
}
\item{ll}{
return value of reversePackageDependencies()
}
}
\description{
recursively call recursivePackageDependencies2() and reversePackageDependencies()
to recursively delete leaf nodes until packs has been depleted to length zero
}
\details{
NOTE that the packages in packs do not need to be loaded or attached to the search() path
}
\examples{
\dontrun{
packs<-c("cardUtils","clickableImageMap","editDriver",
"heartsCIM","iterationDriver","logos","playOneTrick",
"playWholeHandDriverPassParams","probTab","relaxDriver")
master<-PackageDependencies(packs, vector("list",length(packs)),1,TRUE)
}

}
